+++ pykeg/web/context_processors.py
@@ -30,7 +30,6 @@ def kbsite(request):
     if kbsite:
         ret['guest_info']['name'] = kbsite.settings.guest_name
         ret['guest_info']['image'] = kbsite.settings.guest_image
-        ret['SERIAL_NUMBER'] = kbsite.serial_number
         ret['HAVE_SESSIONS'] = models.DrinkingSession.objects.all().count() > 0
         ret['GOOGLE_ANALYTICS_ID'] = kbsite.settings.google_analytics_id